% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{add_empty_pop}
\alias{add_empty_pop}
\title{Modify a dataframe to enable plotting of populations instead of frequencies}
\usage{
add_empty_pop(Muller_df)
}
\arguments{
\item{Muller_df}{Dataframe created by get_Muller_df}
}
\value{
A dataframe that can be used as input in Muller_plot.
}
\description{
The function adds rows at each time point recording the difference between the total population and its maximum value.
Generally there is no need to use this function as Muller_pop_plot calls it automatically.
}
\examples{
Muller_df <- get_Muller_df(example_edges, example_pop_df)
Muller_df2 <- add_empty_pop(Muller_df)

}
\seealso{
\code{\link{get_Muller_df}} \code{\link{Muller_pop_plot}}
}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
