% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{add_root_row}
\alias{add_root_row}
\title{Add a row to the edges list to represent the root node (if not already present).}
\usage{
add_root_row(tree)
}
\arguments{
\item{tree}{Dataframe comprising an adjacency matrix, with column names "Parent" and "Identity"}
}
\value{
The same dataframe including a row representing the root node.
}
\description{
Add a row to the edges list to represent the root node (if not already present).
}
\examples{
tree1 <- data.frame(Parent = c(1,1,1,2,3,4), 
                    Identity = 2:7)
add_root_row(tree1)

}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
