% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{add_start_points}
\alias{add_start_points}
\title{Add rows to a population dataframe to ensure genotype starting points are plotted correctly}
\usage{
add_start_points(pop_df)
}
\arguments{
\item{pop_df}{Dataframe with column names "Generation", "Identity" and "Population"}
}
\value{
The input Dataframe with additional rows.
}
\description{
Add rows to a population dataframe to ensure genotype starting points are plotted correctly
}
\examples{
edges1 <- data.frame(Parent = 1:3, Identity = 2:4)
pop1 <- data.frame(Generation = rep(1:5, each = 4), Identity = rep(1:4, 5), 
                   Population = c(1,0,0,0,1,1,0,0,1,1,1,0,1,1,1,1,1,1,1,1))
add_start_points(pop1)

}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}

