% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplementaryFunctions.R
\name{adj_matrix_to_tree}
\alias{adj_matrix_to_tree}
\title{Create a tree object of class "phylo" from an adjacency matrix}
\usage{
adj_matrix_to_tree(edges)
}
\arguments{
\item{edges}{Dataframe comprising an adjacency matrix, in which the first column is the parent and the second is the daughter.}
}
\value{
A phylo object.
}
\description{
Create a tree object of class "phylo" from an adjacency matrix
}
\examples{
edges1 <- data.frame(Parent = c(1,1,1,3,3), Identity = 2:6)
require(ape)
tree <- adj_matrix_to_tree(edges1)
plot(tree)

}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
