% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_df-data.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Example dataframe}
\format{A dataframe with column names "Generation", "Identity", "Parent", "Population" and "RelativeFitness"}
\usage{
data(example_df)
}
\description{
Example dataframe containing both phylogenetic information and population dynamics.
}
\keyword{datasets}
