% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_edges-data.R
\docType{data}
\name{example_edges}
\alias{example_edges}
\title{Example adjacency matrix}
\format{
A dataframe with column names "Parent" and "Identity"
}
\usage{
data(example_edges)
}
\description{
Example dataframe comprising an adjacency matrix.
}
\keyword{datasets}
