% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pop_df-data.R
\docType{data}
\name{example_pop_df}
\alias{example_pop_df}
\title{Example population dataframe}
\format{
A dataframe with column names "Generation", "Identity" and "Population"
}
\usage{
data(example_pop_df)
}
\description{
Example dataframe containing population dynamics.
}
\keyword{datasets}
