% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{find_start_node}
\alias{find_start_node}
\title{Move to top of adjacency matrix}
\usage{
find_start_node(edges)
}
\arguments{
\item{edges}{Dataframe comprising an adjacency matrix, with column names "Parent" and "Identity"}
}
\value{
The Parent that is the common ancestor.
}
\description{
Returns the Parent value of the common ancestor.
}
\examples{
edges1 <- data.frame(Parent = c(1,1,1,3,3), Identity = 2:6)
find_start_node(edges1)

}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
