% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{get_Adj}
\alias{get_Adj}
\title{Get adjacency list of a tree.}
\usage{
get_Adj(tree)
}
\arguments{
\item{tree}{Dataframe comprising an adjacency matrix, with column names "Parent" and "Identity"}
}
\value{
The adjacency list.
}
\description{
Get adjacency list of a tree.
}
\examples{
tree1 <- data.frame(Parent = c(1,1,1,1,2,3,4), 
                    Identity = 1:7, 
                    Population = c(1, rep(5, 6)))
get_Adj(tree1)

}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
