% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MullerPlot.R
\name{Muller_plot}
\alias{Muller_plot}
\title{Draw a Muller plot using ggplot2}
\usage{
Muller_plot(Muller_df, colour_by = NA, palette = NA, add_legend = FALSE)
}
\arguments{
\item{Muller_df}{Dataframe created by get_Muller_df}

\item{colour_by}{Character containing name of column by which to colour the plot}

\item{palette}{List of colours}

\item{add_legend}{Logical whether to show legend}
}
\value{
None
}
\description{
Draw a Muller plot using ggplot2
}
\examples{
Muller_df <- get_Muller_df(example_edges, example_pop_df, threshold = 0.005)
Muller_plot(Muller_df)

}
\seealso{
\code{\link{get_Muller_df}}
}
\author{
Rob Noble, \email{robjohnnoble@gmail.com}
}
